+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Next
+
+### Breaking Changes
+
+- rename `ResponseErrorCode` to `ResponseError`
+- rename `CreateResponseBase` to `CreateBaseResponse`
+- rename `CreateJsonBaseResponseFunctions` to `CreateBaseJsonResponseFunctions`
+- rename `CreateJsonSuccessResponseFunctions` to `CreateSuccessJsonResponseFunctions`
+- rename `CreateJsonFailureResponseFunctions` to `CreateFailureJsonResponseFunctions`
+
 ## 0.1.0 (2025-12-16)
 
-First release
+initial release